Use consistent size for entity id

+++ entity.cpp
@@ -1,7 +1,7 @@
 #include "entity.hpp"
 #include "world.hpp"
 
-Entity::Entity(World &world, uint16_t index, blit::Point pos, const EntityInfo &info, int rotation) : world(world), index(index), sprite_index(0), info(info) {
+Entity::Entity(World &world, unsigned int index, blit::Point pos, const EntityInfo &info, int rotation) : world(world), index(index), sprite_index(0), info(info) {
     sprite_index = info.base_sprite + rotation;
 
     position = {-16, -16};
